requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==5.0
django-dotenv==1.4.2
gunicorn==21.2.0
idna==3.4
packaging==23.2
psycopg==3.1.14
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
redis[hiredis]
